Create rock-paper-scissors game in console with JS
Used a script tag in an HTML page to create a simple game of Rock Paper Scissors in the console. The user assigns a value to the variable playerSelection and then calls the function playRound(playerSelection,computerSelection) to determine the outcome of the game. User input is case sensitive and must take the form of 'Rock' 'Paper' or 'Scissors'.