Complete 'Revisiting Rock Paper Scissors' with UI
Finished the user interface for the rock paper scissors project. User inputs their selection via three buttons at the top labeled rock, paper, and scissors. The game runs each round and displays the results and current score in an ordered list. The game ends when either the player or the computer reaches 5 points, at which point the end result is displayed below the round list.